requirements.txt
# You could install from this using pip install -r requirements.txt as we always do in the python community
fastapi[standard]
psycopg2-binary